"Fixed all comments, extra lines, optimized imports and change types int, double to not primitive types, added 2 end-points to get bet and transaction history for future work."

+++ LeoProjectKafkaDTOS/src/main/java/org/example/LeoProjectKafkaDTOS/TransactionDTO.java
@@ -9,12 +9,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class TransactionDTO {
 
-    private int transactionId;
+    private Integer transactionId;
 
     private String type;
 
-    private int transactionAmount;
-
-
-
+    private Double transactionAmount;
 }
